PARSE: Implement mode 'S' in ParsePacker
Packing the input data for parse mode 'S' is now implemented through
_pack_scrape_jobs.

The method _prepare_scrape_files is used to search for successful scrape
actions and, if necessary, extract the scrape job ZIP files to get the
downloaded files if cleanup was requested after a scrape job.

_prepare_scrape_files seems too long and complicated. Hopefully, it will
be refactored sometime.