GENERAL: Add convert_2_soup to utils
convert_2_soup function either loads a file into a soup or does nothing
if a soup was already provided.

This will be useful in parsing, so a file will have to be loaded into a soup
only once by the preparser. The loaded soup can then be passed around to
specific parsers without reading the input file again.