DB: Add passed/duplicate columns to commit_jobs and commits
Add "passed" and "duplicate" columns to the CommitJobs and Commits table
interfaces.

Each entry in the "pub" table must have an unique DOI. The "duplicate"
flag is added here to signify that an entry with a DOI already
preexisting in the "pub" table was attempted to be inserted in the
table.

The "passed" flag simply signifies if a commit action resulted in
inserting/overwriting an entry in the "pub" table.

The "job_passes" and "job_duplicates" columns count how many commit
actions passed and how many commit actions contain a "duplicate" DOI.