COMMIT|PARSE: Define the default metadata properties only once
Add the default_metadata.py script to read the default metadata
properties to be parsed from the registered columns of the "parses"
table.

This way, it is ensured that the metadata properties are defined only in
one place (db/tables/registered_tables.py) and are the same across the
codebase.