DB|PARSE: Add normalized date columns to parses, commits and pub tables
Add normalized_received, normalized_accepted and normalized_published
columns to the "parses", "commits" and "pub" tables.

The normalized dates are in ISO format (YYYY-MM-DD) and are best to be
obtained immediately at the parsing stage to ease filtering and
analyzing the final results in the "pub" table.

Obtaining the normalized dates is also implemented for RSC and NAT
journals in this commit to avoid breaking the Parser subclasses.