PARSE: Add publishers_index.json
The "publishers_index.json" file lists the normalized publisher names
and name variants for all publishers for which a parser is implemented.

This data will be utilized by Parser subclasses to obtain normalized
publisher names. This is necessary to ensure that the data in the "pub"
database may be reliably filtered.