PARSE: Implement normalized_publisher() property in parser
Normalizing the publisher name works by matching the name variants
in "publishers_index.json" with the parsed "self.publisher" value
in Parser() subclasses.

The normalized_publisher() property may be implemented in the abstract
Parser class because all the subclasses can use the same algorithm to
find the normalized publisher name.